Added JWT auth
Created auth middleware that validates the auth token from
x-access-token header. User id is encoded in this token and it is
decoded to find the calling user and store them in ctx.state.user.
Created helper functions that can hash and compare passwords as well as
create the JWT token which expires in 1 day.

Added unit tests for password helper functions and auth middleware.