BugFix: Schema $ref incompatible in OpenAPI
Referring to a neighbouring schema file was done using the schema's $id,
however this approach was not compatible with Redoc. Redoc uses
OpenAPI.yml which loads schemas based on their relative path to each other.
However, JSON Schema validator loads schemas based on their $id. Causing
the other to fail.

Fixed by setting top-level $id to be the file name. This allows Redoc to
interpret them as relative locations and validator as IDs.